+++ ratpack-core/src/main/java/ratpack/service/internal/ServicesGraph.java
@@ -121,7 +121,6 @@ public class ServicesGraph {
     startLatch.await();
     StartupFailureException startupFailureException = failureRef.get();
     if (startupFailureException != null) {
-      LOGGER.error("Startup failure", startupFailureException);
       stop(new DefaultEvent(startEvent.getRegistry(), false));
       throw startupFailureException;
     }